DDF-3672 Rework packages
