+++ catalog/ui/catalog-ui-search/src/main/java/org/codice/ddf/catalog/ui/forms/api/FilterVisitor2.java
@@ -11,7 +11,7 @@
  * License is distributed along with this program and can be found at
  * <http://www.gnu.org/licenses/lgpl.html>.
  */
-package org.codice.ddf.catalog.ui.forms.filter;
+package org.codice.ddf.catalog.ui.forms.api;
 
 import java.io.Serializable;
 import java.util.List;
