+++ catalog/ui/catalog-ui-search/src/main/java/org/codice/ddf/catalog/ui/forms/data/QueryTemplateType.java
@@ -21,7 +21,7 @@ import ddf.catalog.data.impl.MetacardTypeImpl;
 import ddf.catalog.data.impl.types.CoreAttributes;
 import ddf.catalog.data.impl.types.SecurityAttributes;
 import java.util.Set;
-import org.codice.ddf.catalog.ui.forms.model.TransformVisitor;
+import org.codice.ddf.catalog.ui.forms.filter.TransformVisitor;
 import org.codice.ddf.catalog.ui.metacard.workspace.QueryMetacardTypeImpl;
 
 /**
