+++ catalog/ui/catalog-ui-search/src/test/java/org/codice/ddf/catalog/ui/forms/filter/VisitableXmlElementImplTest.java
@@ -34,6 +34,7 @@ import net.opengis.filter.v_2_0.PropertyIsLikeType;
 import net.opengis.filter.v_2_0.PropertyIsNilType;
 import net.opengis.filter.v_2_0.PropertyIsNullType;
 import net.opengis.filter.v_2_0.UnaryLogicOpType;
+import org.codice.ddf.catalog.ui.forms.api.FilterVisitor2;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
